New Ryujinx needs an higher value of vm.max_map_count
Since this Ryujunx PR [1] it returns a warning if vm.max_map_count < 524288,
so just set an higher value.

To be coherent with what other Linux distrubitions are doing, use
2147483642.

This is the value used by Batocera 37 itself [2], SteamDeck and Fedora 39 [3].

There is not need to re-set the value each time yuzu or Ryujinx is
started, so just set it once on startup.

[1] https://github.com/Ryujinx/Ryujinx/pull/4702
[2] https://github.com/batocera-linux/batocera.linux/pull/8649
[3] https://fedoraproject.org/wiki/Changes/IncreaseVmMaxMapCount